Links stackview to the form
