+++ lib/assets/javascripts/cartodb3/editor/editor-view.js
@@ -1,19 +1,23 @@
+var cdb = require('cartodb.js');
+var $ = require('jquery');
 var Backbone = require('backbone');
+var AddWidgetsView = require('./add-widgets/add-widgets-view');
 var createTextLabelsTabPane = require('../components/tab-pane/create-text-labels-tab-pane');
 var StackLayoutView = require('../components/stack-layout/stack-layout-view');
 var TabPaneTemplate = require('./tab-pane-template.tpl');
 var EditorWidgetsView = require('./widgets/widgets-view');
-var WidgetsFormContentView = require('./widgets/widgets-form/widgets-form-content-view');
-var cdb = require('cartodb.js');
 
 module.exports = cdb.core.View.extend({
   initialize: function (opts) {
     if (!opts.layerDefinitionsCollection) throw new Error('layerDefinitionsCollection is required');
     if (!opts.widgetDefinitionsCollection) throw new Error('widgetDefinitionsCollection is required');
-    if (!opts.stackLayoutModel) throw new Error('stackLayoutModel is required');
+    if (!opts.modals) throw new Error('modals is required');
+    if (!opts.mapStackLayoutModel) throw new Error('mapStackLayoutModel is required');
 
+    this._modals = opts.modals;
     this._layerDefinitionsCollection = opts.layerDefinitionsCollection;
     this._widgetDefinitionsCollection = opts.widgetDefinitionsCollection;
+    this._mapStackLayoutModel = opts.mapStackLayoutModel;
   },
 
   render: function () {
@@ -24,19 +28,10 @@ module.exports = cdb.core.View.extend({
         return new EditorWidgetsView({
           layerDefinitionsCollection: self._layerDefinitionsCollection,
           widgetDefinitionsCollection: self._widgetDefinitionsCollection,
-          stackLayoutModel: stackLayoutModel
+          stackLayoutModel: self._mapStackLayoutModel
         });
       }
-    }, {
-      createStackView: function (stackLayoutModel, opts) {
-        var widgetDefinitionModel = opts[0];
-        return new WidgetsFormContentView({
-          widgetDefinitionModel: widgetDefinitionModel,
-          layerDefinitionsCollection: self._layerDefinitionsCollection,
-          stackLayoutModel: stackLayoutModel
-        });
-      }
-    }]);
+    } ]);
 
     var tabPaneTabs = [{
       label: _t('editor.layers.title'),
@@ -76,6 +71,20 @@ module.exports = cdb.core.View.extend({
 
     this._mapTabPaneView = createTextLabelsTabPane(tabPaneTabs, tabPaneOptions);
     this.$el.append(this._mapTabPaneView.render().$el);
+
+    // TODO tmp; move to edit-content-view once the top-level tab pane is implemented
+    var $addWidgetsBtn = $('<button class="CDB-Button CDB-Button--primary"><span>Add widgets</span></button>');
+    this.$el.append($addWidgetsBtn);
+    $addWidgetsBtn.on('click', function () {
+      // Open a add-widgets-modal on page load
+      self._modals.create(function (modalModel) {
+        return new AddWidgetsView({
+          modalModel: modalModel,
+          layerDefinitionsCollection: self._layerDefinitionsCollection,
+          widgetDefinitionsCollection: self._widgetDefinitionsCollection
+        });
+      });
+    });
 
     return this;
   }
