+++ lib/assets/javascripts/cartodb3/editor/map/map-view.js
@@ -1,10 +1,9 @@
-var $ = require('jquery');
 var Backbone = require('backbone');
 var cdb = require('cartodb-deep-insights.js');
 var EditorView = require('../editor-view');
 var Template = require('./template.tpl');
-var AddWidgetsView = require('../add-widgets/add-widgets-view');
 var StackLayoutView = require('../../components/stack-layout/stack-layout-view');
+var WidgetsFormContentView = require('../widgets/widgets-form/widgets-form-content-view');
 
 module.exports = cdb.core.View.extend({
   initialize: function (opts) {
@@ -29,8 +28,18 @@ module.exports = cdb.core.View.extend({
     var stackViewCollection = new Backbone.Collection([{
       createStackView: function (stackLayoutModel, opts) {
         return new EditorView({
+          modals: self._modals,
           layerDefinitionsCollection: self._layerDefinitionsCollection,
           widgetDefinitionsCollection: self._widgetDefinitionsCollection,
+          stackLayoutModel: stackLayoutModel
+        });
+      }
+    }, {
+      createStackView: function (stackLayoutModel, opts) {
+        var widgetDefinitionModel = opts[0];
+        return new WidgetsFormContentView({
+          widgetDefinitionModel: widgetDefinitionModel,
+          layerDefinitionsCollection: self._layerDefinitionsCollection,
           stackLayoutModel: stackLayoutModel
         });
       }
@@ -41,20 +50,6 @@ module.exports = cdb.core.View.extend({
     });
 
     this.$('.js-content').append(stackLayoutView.render().$el);
-
-    // TODO tmp; move to edit-content-view once the top-level tab pane is implemented
-    var $addWidgetsBtn = $('<button class="CDB-Button CDB-Button--primary"><span>Add widgets</span></button>');
-    stackLayoutView.$el.append($addWidgetsBtn);
-    $addWidgetsBtn.on('click', function () {
-      // Open a add-widgets-modal on page load
-      self._modals.create(function (modalModel) {
-        return new AddWidgetsView({
-          modalModel: modalModel,
-          layerDefinitionsCollection: self._layerDefinitionsCollection,
-          widgetDefinitionsCollection: self._widgetDefinitionsCollection
-        });
-      });
-    });
 
     return this;
   }
